Move Hungarian matching utilities to nn.utils
