+++ sleap/nn/utils.py
@@ -1,8 +1,10 @@
 """This module contains generic utilities used for training and inference."""
 
+import tensorflow as tf
 import numpy as np
 from collections import defaultdict
-from typing import Dict
+from typing import Dict, Tuple
+from scipy.optimize import linear_sum_assignment
 
 
 def group_array(
@@ -70,3 +72,55 @@ def compute_iou(bbox1: np.ndarray, bbox2: np.ndarray) -> float:
     iou = intersection_area / union_area
 
     return iou
+
+
+@tf.function
+def tf_linear_sum_assignment(cost_matrix: tf.Tensor) -> tf.Tensor:
+    """Run `linear_sum_assignment` as a TensorFlow function.
+
+    Args:
+        cost_matrix: Cost matrix of shape `(n_src, n_dst)`. Make sure to replace `NaN`s
+            with `np.inf`.
+
+    Returns:
+        A tuple of `(row, col)` with the indices of the optimal assignments.
+    """
+
+    def _linear_sum_assignment(cost_matrix):
+        """Wrap `linear_sum_assignment` for type safety."""
+        row, col = linear_sum_assignment(cost_matrix)
+        return row.astype("int32"), col.astype("int32")
+
+    return tf.numpy_function(
+        func=_linear_sum_assignment, inp=[cost_matrix], Tout=[tf.int32, tf.int32]
+    )
+
+
+def match_points(points1: tf.Tensor, points2: tf.Tensor) -> Tuple[tf.Tensor, tf.Tensor]:
+    """Match closest points across two sets.
+
+    Args:
+        points1: A set of reference points of shape `(N, D)`.
+        points2: A set of unmatched points of shape `(M, D)`.
+
+    Returns:
+        A tuple of `(inds1, inds2)` with the indices of matching points across the two
+        lists of points.
+
+        `inds1` is a `tf.int32` tensor of shape `(n,)` where `n <= N` that indexes into
+        `points1`.
+
+        `inds2` is a `tf.int32` tensor of shape `(m,)` where `m <= M` that indexes into
+        `points2`.
+
+    Notes:
+        This uses Euclidean distance and the Hungarian algorithm for matching.
+
+    See also: tf_linear_sum_assignment
+    """
+    dists = tf.linalg.norm(
+        tf.expand_dims(tf.cast(points1, tf.float32), axis=1)
+        - tf.expand_dims(tf.cast(points2, tf.float32), axis=0),
+        axis=-1,
+    )
+    return tf_linear_sum_assignment(dists)
